Refactored all views to extand AbstractView from framework
