+++ src/main.js
@@ -1,9 +1,10 @@
 import { Route, Point, Offers } from './model';
 import { RoutePresenter } from './presenter';
-import { render } from './render';
+import { render } from './framework/render';
 import { AddPointButtonView, FiltersView } from './view';
 import { generateOffers } from './mock/generate-offers';
 import { generateDestinations } from './mock/generate-destinations';
+import { generatePoints } from './mock/generate-points';
 
 const tripContainer = document.querySelector('.trip-main');
 const routeContainer = document.querySelector('.trip-events');
@@ -14,7 +15,7 @@ const offersModel = Offers.parse(generateOffers());
 const destinations = generateDestinations();
 const routePresenter = new RoutePresenter(routeContainer, routeModel, offersModel, destinations);
 
-routeModel.points = Point.parseAll([]); // TODO: make same as offers parse
+routeModel.points = Point.parseAll(generatePoints()); // TODO: make same as offers parse
 
 render(filtersView, tripContainer);
 render(addPointButtonView, tripContainer);
