+++ src/view/add-point-button/add-point-button.js
@@ -1,4 +1,4 @@
-import AbstractView from '../abstract-view/abstract-view';
+import AbstractView from '../../framework/view/abstract-view';
 import { createAddPointButtonTemplate } from './create-add-point-button-template';
 
 export default class AddPointButtonView extends AbstractView {
@@ -7,7 +7,7 @@ export default class AddPointButtonView extends AbstractView {
    * Returns view's html template
    * @returns {String} - view's template
    */
-  getTemplate() {
+  get template() {
     return createAddPointButtonTemplate();
   }
 }
