+++ src/view/add-point/add-point.js
@@ -1,4 +1,4 @@
-import AbstractView from '../abstract-view/abstract-view';
+import AbstractView from '../../framework/view/abstract-view';
 import { createAddPointTemplate } from './create-add-point-template';
 
 export default class AddPointView extends AbstractView {
@@ -7,7 +7,7 @@ export default class AddPointView extends AbstractView {
    * Returns view's html template
    * @returns {String} - view's template
    */
-  getTemplate() {
+  get template() {
     return createAddPointTemplate();
   }
 }
