+++ src/view/filters/filters.js
@@ -1,4 +1,4 @@
-import AbstractView from '../abstract-view/abstract-view';
+import AbstractView from '../../framework/view/abstract-view';
 import { createFiltersTemplate } from './create-filters-template';
 
 export default class FiltersView extends AbstractView {
@@ -7,7 +7,7 @@ export default class FiltersView extends AbstractView {
    * Returns view's html template
    * @returns {String} - view's template
    */
-  getTemplate() {
+  get template() {
     return createFiltersTemplate();
   }
 }
