+++ src/view/no-points/no-points.js
@@ -1,4 +1,4 @@
-import AbstractView from '../abstract-view/abstract-view';
+import AbstractView from '../../framework/view/abstract-view';
 import { createNoPointsTemplate } from './create-no-points-template';
 
 export default class NoPointsView extends AbstractView {
@@ -18,7 +18,7 @@ export default class NoPointsView extends AbstractView {
    * Returns view's html template
    * @returns {String} - view's template
    */
-  getTemplate() {
+  get template() {
     return createNoPointsTemplate(this.#message);
   }
 }
