+++ src/view/point-list/point-list.js
@@ -1,4 +1,4 @@
-import AbstractView from '../abstract-view/abstract-view';
+import AbstractView from '../../framework/view/abstract-view';
 import { createPointListTemplate } from './create-point-list-template';
 
 export default class PointListView extends AbstractView {
@@ -7,7 +7,7 @@ export default class PointListView extends AbstractView {
    * Returns view's html template
    * @returns {String} - view's template
    */
-  getTemplate() {
+  get template() {
     return createPointListTemplate();
   }
 }
